Applied patch from Andrew Feldman:
Hi Nathanael,

I have been using image resizer for some time on a previous project with great success.  Thank you so much for this wonderful utility.  I am incorporating the utility into my latest project and I have found one issue in the latest bits, though it was probably there before, I was just able to work around it as it is only a problem the first time the app is run in a new location.  The issue lies in the DiskCache.PrepareCacheDir method.

The method checks to see whether a web.config exists in the folder, and if not it will create one.  There appears to be a concurrency issue here.  If I am displaying a list of 15 images in a list that need resizing, the first request will trigger the write to the file system, but the second request will puke if the first has not finished.  I have added a lock and a double check for the existance of the web.config to save additional file system hits on subsequent requests which may be waiting on the lock.  I am no longer able to repro the issue with this change.  Here is what the change looks like... I can't get formatting to work properly in here, so you'll want to paste it in Visual Studio and format it.

if

(!wc.FileExists)
{

lock (SyncRoot)
{

if (!wc.FileExists)
{

System.IO.

File.WriteAllText(wc.Local, webConfigFile);
}

}

}

Where SyncRoot is declared private static readonly object SyncRoot = new object();

Thanks,
Andrew