Fixed bug reported by adam duke. Cached files modified by just one day or one hour don't get updated.
in the RoughCompare function it is currently using Timespan.Milliseconds, which only takes the millisecond component of the time span.  Thus, if somebody updates a file exactly 1 day and 1 hour, the comparison yields a big fat zero and the cache is never updated.

If you change the comparison to use TotalMilliseconds instead, then everything works as expected.  Might be worth rolling back into the next update (if you haven't found it already).


Former-commit-id: 324dff72beebd666d2a9629dbe1d5a16c99a3f43