Added support for app-relative destination paths to ImageBuilder.Build.
Modified IEncoder and IEncoderProvider interface. ResizeSettings instance is now first (due to it being more important), and the second 'original' paramater is of type object, allowing paths to be passed directly.
DefaultEncoder supports path parsing for the 'original' param, and by consequence, PrettyGifs does also.

This makes it much easier for users to acquire an encoder instance for purposes of extension and mime-type determination.


Former-commit-id: fe8cee805f2913cf681c5f94f6cff4b72eb9752e