Added first draft of 'cropxunits' and 'cropyunits'.
These can be set to a decimal value.
If set to a decimal value, crop units for that dimenision will be interpreted as relative to the value. This allows easy cropping without knowing the original size of the image.

For example, ?crop=10,10,90,90&cropxunits=100&cropyunits=100 will crop a 10% border off each edge.