Fixes issue #26: Created `SeekableStreamWrapper` that can ensure a source stream is wrapped by a seekable helper if it isn't already seekable.
Implemented in Core/Util so that (a) it's available to any code with a reference to ImageResizer, (b) it was easy to put tests into Core.Tests, and (c) future enhancements (buffer on-demand, instead of all up-front) can be done in-place without altering calling code.  For now, only RemoteReader is making use of the class, but since it's in Core, it can also be used by the ImageResizer framework if desired.