Fixes #69: pre-fetch watermark images into the ASP.NET cache _before_ the main image is loaded.
The pre-fetch behavior is supported for named image watermarks (`<image>` in the config), explicit watermark files (a path in the querystring's `watermark` value, and an `<otherimages>` setting in the config), and also legacy watermarks (a path in the querystring's `watermark` value, but _no_ `<otherimages>` setting in the config).  Gradients are _not_ pre-fetched, as their load-size depends on the size needed for the image on which they are applied.


Former-commit-id: 47a0ad2e5f36c8550942b41c7b54248d2521d69c