add the ability to have certain query string parameters excluded from the cache key, etc.
sometimes one might want to exclude certain query string parameters from
the cache key, e.g. when using a qs parameter as a cache buster for a
CDN, etc.

added the ability to do this via a config item:

<ignoredquerystringparameters parameternames="" />

the parameternames attribute accepts a comma-delimited list of parameter
names that will be ignored by internal processing (and thus also from
the cache key)