PdfiumRenderer: Move native dependency in test configuration for compatibility with TestDriven.NET.
TestDriven.NET does not run from the output directory but instead creates a copy of the output directory to a temporary folder. However, only the contents of the root directory is copied. To fix this the x86 version of the native dependency is now output to the root of the output directory, only for the test project.